create allocator module
