+++ python/sglang/srt/mem_cache/unified_cache_components/swa_component.py
@@ -39,7 +39,7 @@ class SWAComponent(TreeComponent):
     """
 
     def __init__(self, cache: UnifiedRadixCache, params: CacheInitParams):
-        from sglang.srt.mem_cache.swa_memory_pool import SWATokenToKVPoolAllocator
+        from sglang.srt.mem_cache.allocator.swa import SWATokenToKVPoolAllocator
 
         assert isinstance(
             cache.token_to_kv_pool_allocator, SWATokenToKVPoolAllocator
